- Changed class structure. Renamed Datasource.java to Database.java -- Removed Datasource.java
